refactor(server): Extract build logic into separate module
This module is going to get more complex as the implementation of #32
progresses.
